Implement function to find the nodecg installation by ourselfs (#38)
Currently we're using [find-up](https://www.npmjs.com/package/find-up) to traverse the filesystem and find the nodecg installation if the user is an subdirectory like `bundles/`.
To reduce our dependencies we can simply implement this ourselfs. I can't remember why I used `find-up` in the first place as using it is almost harder than implementing it directly, lol.