refactor(fs): Replace custom directory delete implementation with node fs method
With ff8808a742e4936fe6f4974c151c12b9aa1d6f80 we switched the cli to
node.js 14.14.0 or higher. With this version the `recursive` option was
added to the rm method in the node.js fs module.
Now we can use this integrated method as we don't need to support
the older node.js versions anymore.